backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
httpx==0.25.1
python-dotenv==1.0.0
pydantic==2.4.2
pydantic-settings==2.0.3
python-multipart==0.0.6
aiofiles==23.2.1
wordcloud==1.9.3
numpy==1.26.2
Pillow==10.2.0
matplotlib==3.8.3
requests==2.31.0
pydantic-core==2.10.1
sniffio==1.3.1
starlette==0.27.0
typing-extensions==4.14.0
urllib3==2.4.0
nltk==3.8.1
pandas==2.1.3
scikit-learn==1.3.2
sse-starlette==1.6.5

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Yandex Cloud
yandexcloud==0.227.0 